moved urlRegex to common and updated references
Jira ticket: CAMS-418
