+++ backend/functions/lib/adapters/gateways/user-session-gateway.test.ts
@@ -13,7 +13,7 @@ import * as factoryModule from '../../factory';
 import { ServerConfigError } from '../../common-errors/server-config-error';
 import { CamsSession } from '../../../../../common/src/cams/session';
 import { CamsRole } from '../../../../../common/src/cams/roles';
-import { urlRegex } from '../../../../../user-interface/src/lib/testing/testing-utilities';
+import { urlRegex } from '../../../../../common/src/cams/test-utilities/auxiliary';
 import { OFFICES } from '../../../../../common/src/cams/test-utilities/offices.mock';
 import { CamsJwtHeader } from '../../../../../common/src/cams/jwt';
