+++ common/src/cams/test-utilities/auxiliary.ts
@@ -1,0 +1,1 @@
+export const urlRegex = /https?:\/\/.*\//;
