+++ user-interface/src/login/providers/okta/okta-library.test.ts
@@ -4,8 +4,9 @@ import { getCamsUser, refreshOktaToken, registerRefreshOktaToken } from './okta-
 import LocalStorage from '@/lib/utils/local-storage';
 import { MockData } from '@common/cams/test-utilities/mock-data';
 import * as apiModule from '@/lib/models/api';
-import TestingUtilities, { urlRegex } from '@/lib/testing/testing-utilities';
+import TestingUtilities from '@/lib/testing/testing-utilities';
 import { CamsSession } from '@common/cams/session';
+import { urlRegex } from '../../../../../common/src/cams/test-utilities/auxiliary';
 
 const MOCK_OAUTH_CONFIG = { issuer: 'https://mock.okta.com/oauth2/default' };
